move legacy Plugin base class from core to config
