+++ rdmo/config/plugins.py
@@ -4,6 +4,11 @@ from django.conf import settings
 
 from rdmo.config.helpers import DeclaredPlugin
 from rdmo.config.models import Plugin
+from rdmo.core.utils import import_class
+
+
+class PluginBase:
+    pass
 
 
 def get_plugins_from_db(project=None, plugin_type: str | None=None, user=None) -> list[DeclaredPlugin]:
@@ -29,7 +34,7 @@ def get_plugins_from_db(project=None, plugin_type: str | None=None, user=None) -
             continue
 
         if instance.python_path not in settings.PLUGINS:
-            continue
+            continue  # plugins are filtered by settings as well
 
         if instance.plugin_type != plugin_type:
             continue
@@ -44,3 +49,15 @@ def get_plugins_from_db(project=None, plugin_type: str | None=None, user=None) -
             source=f"Plugin(id={instance.id})",
         ))
     return results
+
+
+def get_plugin(plugin_settings, plugin_key):
+    try:
+        key, label, class_name = next(
+            (key, label, class_name)
+            for key, label, class_name in getattr(settings, plugin_settings)
+            if key == plugin_key
+        )
+        return import_class(class_name)(key, label, class_name)
+    except StopIteration:
+        return None
