+++ rdmo/projects/imports.py
@@ -11,8 +11,8 @@ from django.utils.translation import gettext_lazy as _
 
 import requests
 
+from rdmo.config.plugins import PluginBase
 from rdmo.core.imports import handle_fetched_file
-from rdmo.core.plugins import Plugin
 from rdmo.core.xml import get_ns_map, get_uri, read_xml_file
 from rdmo.domain.models import Attribute
 from rdmo.options.models import Option
@@ -25,7 +25,7 @@ from .models import Project, Snapshot, Value
 log = logging.getLogger(__name__)
 
 
-class Import(Plugin):
+class Import(PluginBase):
 
     plugin_type = 'project_import'
     accept = None
